V 2.8.12 Added generic function for lang.rs
