+++ json/message_command/test.json
@@ -1,3 +1,2 @@
 {
-
 }
