+++ src/structure/message/admin/anilist/add_activity.rs
@@ -6,6 +6,7 @@ use serde::{Deserialize, Serialize};
 use crate::helper::error_management::error_enum::{AppError, ErrorResponseType, ErrorType};
 use crate::helper::get_guild_lang::get_guild_language;
 use crate::helper::read_file::read_file_as_string;
+use crate::structure::message::common::load_localization;
 
 /// `AddActivityLocalised` is a struct that represents an add activity's localized data.
 /// It contains four fields `success`, `fail`, `fail_desc` and `success_desc` which are all Strings.
@@ -43,23 +44,5 @@ pub async fn load_localization_add_activity(
     guild_id: String,
 ) -> Result<AddActivityLocalised, AppError> {
     let path = "json/message/admin/anilist/add_activity.json";
-    let json = read_file_as_string(path)?;
-    // Parse the JSON data into a HashMap and handle any potential errors
-    let json_data: HashMap<String, AddActivityLocalised> =
-        serde_json::from_str(&json).map_err(|e| {
-            AppError::new(
-                format!("Failing to parse add_activity.json. {}", e),
-                ErrorType::File,
-                ErrorResponseType::Unknown,
-            )
-        })?;
-
-    // Get the language choice for the guild
-    let lang_choice = get_guild_language(guild_id).await;
-
-    // Retrieve the localized data for the add activity based on the language choice
-    Ok(json_data
-        .get(lang_choice.as_str())
-        .cloned()
-        .unwrap_or(json_data.get("en").unwrap().clone()))
+    load_localization(guild_id, path).await
 }
