+++ src/structure/message/admin/module.rs
@@ -7,6 +7,7 @@ use serde_json::from_str;
 use crate::helper::error_management::error_enum::{AppError, ErrorResponseType, ErrorType};
 use crate::helper::get_guild_lang::get_guild_language;
 use crate::helper::read_file::read_file_as_string;
+use crate::structure::message::common::load_localization;
 
 /// `ModuleLocalised` is a struct that represents a module's localized data.
 /// It contains two fields `on` and `off` which are both Strings.
@@ -40,22 +41,6 @@ pub async fn load_localization_module_activation(
     guild_id: String,
 ) -> Result<ModuleLocalised, AppError> {
     let path = "json/message/admin/module.json";
-    let json = read_file_as_string(path)?;
-    // Parse the JSON string into a HashMap.
-    let json_data: HashMap<String, ModuleLocalised> = from_str(&json).map_err(|e| {
-        AppError::new(
-            format!("Failing to parse module.json. {}", e),
-            ErrorType::File,
-            ErrorResponseType::Unknown,
-        )
-    })?;
+    load_localization(guild_id, path).await
 
-    // Get the language choice based on the guild_id.
-    let lang_choice = get_guild_language(guild_id).await;
-
-    // Return the localized data for the module or an error if the language is not found.
-    Ok(json_data
-        .get(lang_choice.as_str())
-        .cloned()
-        .unwrap_or(json_data.get("en").unwrap().clone()))
 }
