+++ src/structure/message/anilist_user/media.rs
@@ -6,6 +6,7 @@ use serde::{Deserialize, Serialize};
 use crate::helper::error_management::error_enum::{AppError, ErrorResponseType, ErrorType};
 use crate::helper::get_guild_lang::get_guild_language;
 use crate::helper::read_file::read_file_as_string;
+use crate::structure::message::common::load_localization;
 
 /// MediaLocalised struct represents a media's localized data.
 /// It contains fields for two titles, a description, and staff text.
@@ -40,22 +41,6 @@ pub struct MediaLocalised {
 /// This function will return an error if the JSON file cannot be read, the JSON cannot be parsed, or the language is not found.
 pub async fn load_localization_media(guild_id: String) -> Result<MediaLocalised, AppError> {
     let path = "json/message/anilist_user/media.json";
-    let json = read_file_as_string(path)?;
-    // Parse the JSON data into a HashMap and handle any potential errors
-    let json_data: HashMap<String, MediaLocalised> = serde_json::from_str(&json).map_err(|e| {
-        AppError::new(
-            format!("Failing to parse media.json. {}", e),
-            ErrorType::File,
-            ErrorResponseType::Unknown,
-        )
-    })?;
+    load_localization(guild_id, path).await
 
-    // Get the language choice for the guild
-    let lang_choice = get_guild_language(guild_id).await;
-
-    // Retrieve the localized data for the media item based on the language choice
-    Ok(json_data
-        .get(lang_choice.as_str())
-        .cloned()
-        .unwrap_or(json_data.get("en").unwrap().clone()))
 }
