+++ src/structure/message/anilist_user/user.rs
@@ -6,6 +6,7 @@ use serde::{Deserialize, Serialize};
 use crate::helper::error_management::error_enum::{AppError, ErrorResponseType, ErrorType};
 use crate::helper::get_guild_lang::get_guild_language;
 use crate::helper::read_file::read_file_as_string;
+use crate::structure::message::common::load_localization;
 
 /// UserLocalised struct represents a user's localized data.
 /// It contains fields for manga, anime, week, day, hour, minute, weeks, days, hours, and minutes.
@@ -48,22 +49,6 @@ pub struct UserLocalised {
 /// * `Result<UserLocalised, AppError>`: A Result containing UserLocalised data or an AppError.
 pub async fn load_localization_user(guild_id: String) -> Result<UserLocalised, AppError> {
     let path = "json/message/anilist_user/user.json";
-    let json = read_file_as_string(path)?;
-    // Parse the JSON data into a HashMap and handle any potential errors
-    let json_data: HashMap<String, UserLocalised> = serde_json::from_str(&json).map_err(|e| {
-        AppError::new(
-            format!("Failing to parse user.json. {}", e),
-            ErrorType::File,
-            ErrorResponseType::Unknown,
-        )
-    })?;
+    load_localization(guild_id, path).await
 
-    // Get the language choice for the guild
-    let lang_choice = get_guild_language(guild_id).await;
-
-    // Return the localized data for the language or an error if the language is not found.
-    Ok(json_data
-        .get(lang_choice.as_str())
-        .cloned()
-        .unwrap_or(json_data.get("en").unwrap().clone()))
 }
