+++ src/structure/message/anime_nsfw/random_image_nsfw.rs
@@ -5,6 +5,7 @@ use serde::{Deserialize, Serialize};
 use crate::helper::error_management::error_enum::{AppError, ErrorResponseType, ErrorType};
 use crate::helper::get_guild_lang::get_guild_language;
 use crate::helper::read_file::read_file_as_string;
+use crate::structure::message::common::load_localization;
 
 /// Represents a random NSFW image's localized data.
 ///
@@ -37,23 +38,6 @@ pub async fn load_localization_random_image_nsfw(
     guild_id: String,
 ) -> Result<RandomImageNSFWLocalised, AppError> {
     let path = "json/message/anime_nsfw/random_image_nsfw.json";
-    let json = read_file_as_string(path)?;
-    // Parse the JSON string into a HashMap.
-    let json_data: HashMap<String, RandomImageNSFWLocalised> = serde_json::from_str(&json)
-        .map_err(|e| {
-            AppError::new(
-                format!("Failing to parse random_image_nsfw.json. {}", e),
-                ErrorType::File,
-                ErrorResponseType::Unknown,
-            )
-        })?;
+    load_localization(guild_id, path).await
 
-    // Get the language choice based on the guild_id.
-    let lang_choice = get_guild_language(guild_id).await;
-
-    // Return the localized data for the language or an error if the language is not found.
-    Ok(json_data
-        .get(lang_choice.as_str())
-        .cloned()
-        .unwrap_or(json_data.get("en").unwrap().clone()))
 }
