+++ src/structure/message/game/steam_game_info.rs
@@ -5,6 +5,7 @@ use serde::{Deserialize, Serialize};
 use crate::helper::error_management::error_enum::{AppError, ErrorResponseType, ErrorType};
 use crate::helper::get_guild_lang::get_guild_language;
 use crate::helper::read_file::read_file_as_string;
+use crate::structure::message::common::load_localization;
 
 /// `SteamGameInfoLocalised` is a struct that represents a Steam game's localized data.
 /// It contains several fields which are all Strings.
@@ -45,23 +46,6 @@ pub async fn load_localization_steam_game_info(
     guild_id: String,
 ) -> Result<SteamGameInfoLocalised, AppError> {
     let path = "json/message/game/steam_game_info.json";
-    let json = read_file_as_string(path)?;
-    // Parse the JSON data into a HashMap and handle any potential errors
-    let json_data: std::collections::HashMap<String, SteamGameInfoLocalised> =
-        serde_json::from_str(&json).map_err(|e| {
-            AppError::new(
-                format!("Failing to parse steam_game_info.json. {}", e),
-                ErrorType::File,
-                ErrorResponseType::Unknown,
-            )
-        })?;
+    load_localization(guild_id, path).await
 
-    // Get the language choice for the guild
-    let lang_choice = get_guild_language(guild_id).await;
-
-    // Return the localized data for the language or an error if the language is not found.
-    Ok(json_data
-        .get(lang_choice.as_str())
-        .cloned()
-        .unwrap_or(json_data.get("en").unwrap().clone()))
 }
