+++ src/structure/message/new_member.rs
@@ -6,6 +6,7 @@ use serde_json::from_str;
 use crate::helper::error_management::error_enum::{AppError, ErrorResponseType, ErrorType};
 use crate::helper::get_guild_lang::get_guild_language;
 use crate::helper::read_file::read_file_as_string;
+use crate::structure::message::common::load_localization;
 
 /// Represents the localized messages for a new member.
 ///
@@ -33,22 +34,6 @@ pub async fn load_localization_new_member(
     guild_id: String,
 ) -> Result<NewMemberLocalised, AppError> {
     let path = "json/message/new_member.json";
-    let json = read_file_as_string(path)?;
-    // Parse the JSON string into a HashMap.
-    let json_data: HashMap<String, NewMemberLocalised> = from_str(&json).map_err(|e| {
-        AppError::new(
-            format!("Error parsing new_member.json. {}", e),
-            ErrorType::File,
-            ErrorResponseType::Unknown,
-        )
-    })?;
+    load_localization(guild_id, path).await
 
-    // Get the language choice based on the guild_id.
-    let lang_choice = get_guild_language(guild_id).await;
-
-    // Return the localized data for the language or an error if the language is not found.
-    Ok(json_data
-        .get(lang_choice.as_str())
-        .cloned()
-        .unwrap_or(json_data.get("en").unwrap().clone()))
 }
